+++ downloader/downloader.go
@@ -1,4 +1,4 @@
-package main
+package downloader
 
 import (
 	"fmt"
