+++ scanner/scanner.go
@@ -1,4 +1,3 @@
-// scanner/scanner.go
 package scanner
 
 import (
